Restructuring of library
