+++ .gitignore
@@ -23,3 +23,4 @@ cabal.project.local~
 .ghc.environment.*
 .vscode
 .DS_Store
+*.wav
