+++ src/Synthesizer.hs
@@ -1,0 +1,2 @@
+module Synthesizer where
+
